Add context variable and remove params
SiteInfo and Web both may need access to the Context object so the constructor saves a reference in a private/protected variable. Means parameter can be removed from some functions. One thought though is that it could be passed in as a parameter to the constructor ratehr than calling the singleton get function. Need to think about that.